improved lockRow and added helper function to loadTiles
LockRow will now look at valid moves and will first check if the row is already locked, if it is we will see if any moves will increase the adj tile value. If not, we will check the resulting valid moves to see if any are better positions than previous. Lock row and Load tiles will always return a move, even if there is only one valid move to make.