Created a scoring system and point system to evaluate moves
Created four function. One in grid to return an array of maxTiles. A function that creates an array of valid moves. A function that goes through a move for a given state and determines the points for that moved based on a combination of largest in the corner, descending order, loading tiles, locking tiles, and maintaining the structure. There is also a function that evaluates the points for every valid move and returns the move with the best score. These functions look at one resulting grid in the future and return the "best " move based off our score metrics.