Add buttons and extra functionality to GOL, limited designer size
Make functionality to play life with 2048 boards.
Creates sucessive board states of GOL and checks to see if they are makeable with the current image creation method.
Use a global timeout variable to start and stop life boards from generating, using the "start life" and "stop life" buttons
Allow for different change of speed for GOL
Limit the designer to size 8 for now, designer goes away if size is > 8